changed long double to double

+++ FFT.cpp
@@ -8,8 +8,8 @@
 #include <complex> // to use complex numbers
 #include <cmath>
 
-typedef long double num;
-typedef std::vector<long double>::const_iterator NumIter;
+typedef double num;
+typedef std::vector<double>::const_iterator NumIter;
 
 typedef std::complex<num> complex_num;
